Add experimental failgrind tool.
This tool simulates heap memory and syscall failures by always failing the first time it sees an allocation or syscall request for a particular call stack. It stores information on what call stacks have gone before so it can run multiple times for the same program and test different parts of the code.